Changed semantics of prefetch-threshold, now a low-mark

+++ src/main/java/no/ssb/dc/api/node/Paginate.java
@@ -13,7 +13,7 @@ public interface Paginate extends NodeWithId {
 
     List<Execute> targets();
 
-    double threshold();
+    int threshold();
 
     Condition condition();
